migrate msg customizations to class override mechanism.
