+++ temba/urls.py
@@ -1,7 +1,6 @@
 from django.conf import settings
 from django.conf.urls import include, url
 from django.conf.urls.static import static
-from django.contrib.auth.models import AnonymousUser, User
 from django.views.generic import RedirectView
 from django.views.i18n import JavaScriptCatalog
 
@@ -12,6 +11,10 @@ from temba.utils.analytics import init_analytics
 
 # javascript translation packages
 js_info_dict = {"packages": ()}  # this is empty due to the fact that all translation are in one folder
+
+# import our surgical overrides that required all Django apps and __init__.pys processed.
+from engage.utils.overrides import EngageOverrides
+EngageOverrides.RunEngageOverrides()
 
 urlpatterns = [
     url(r"^", include("temba.airtime.urls")),
@@ -59,22 +62,6 @@ init_analytics()
 @worker_process_init.connect
 def configure_workers(sender=None, **kwargs):
     init_analytics()  # pragma: needs cover
-
-
-def track_user(self):  # pragma: no cover
-    """
-    Should the current user be tracked
-    """
-
-    # nothing to report if they haven't logged in
-    if not self.is_authenticated:
-        return False
-
-    return True
-
-
-User.track_user = track_user
-AnonymousUser.track_user = track_user
 
 
 def handler500(request):
